% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_status}
\alias{job_status}
\title{Get the current status of a run}
\usage{
job_status(job, refresh = TRUE, config = NULL)
}
\arguments{
\item{job}{A \code{run_id} or \code{laims_dgx_job} object.}

\item{refresh}{Should the function sync against the remote sources first?}

\item{config}{A \code{laims_dgx_config} object.}
}
\value{
An object of class \code{laims_dgx_status}.
}
\description{
Get the current status of a run
}
