% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{recover_job}
\alias{recover_job}
\title{Recover a job handle from the local registry}
\usage{
recover_job(run_id, config = NULL)
}
\arguments{
\item{run_id}{Canonical run identifier.}

\item{config}{A \code{laims_dgx_config} object.}
}
\value{
An object of class \code{laims_dgx_job}.
}
\description{
Recover a job handle from the local registry
}
