% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{sync_job}
\alias{sync_job}
\title{Synchronize one run with remote status sources}
\usage{
sync_job(job, config = NULL)
}
\arguments{
\item{job}{A \code{run_id} or \code{laims_dgx_job} object.}

\item{config}{A \code{laims_dgx_config} object.}
}
\value{
A refreshed \code{laims_dgx_job} object.
}
\description{
Synchronize one run with remote status sources
}
